pyproject.toml
[project]
name = "agencymanager"
version = "0.1.0"
description = "Insurance Agency Management PoC with React frontend and agentic backend"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.116.0",
    "openai>=1.99.0",
    "pydantic-ai>=0.8.0",
    "pydantic-settings>=2.10.0",
    "uvicorn[standard]>=0.35.0",
]
